+++ lib/pbio/include/pbsys/bluetooth.h
@@ -33,6 +33,13 @@ pbio_error_t pbsys_bluetooth_rx(uint8_t *data, uint32_t *size);
 pbio_error_t pbsys_bluetooth_tx(const uint8_t *data, uint32_t *size);
 bool pbsys_bluetooth_tx_is_idle(void);
 
+#if PBSYS_CONFIG_BLUETOOTH_TOGGLE
+void pbsys_bluetooth_enabled_state_request_toggle(void);
+#else
+static inline void pbsys_bluetooth_enabled_state_request_toggle(void) {
+}
+#endif
+
 #else // PBSYS_CONFIG_BLUETOOTH
 
 #define pbsys_bluetooth_init()
@@ -48,6 +55,8 @@ static inline pbio_error_t pbsys_bluetooth_tx(const uint8_t *data, uint32_t *siz
 }
 static inline bool pbsys_bluetooth_tx_is_idle(void) {
     return false;
+}
+static inline void pbsys_bluetooth_enabled_state_request_toggle(void) {
 }
 
 #endif // PBSYS_CONFIG_BLUETOOTH