+++ lib/pbio/platform/prime_hub/pbsysconfig.h
@@ -1,8 +1,12 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2020-2023 The Pybricks Authors
 
+#include <pbio/button.h>
+
 #define PBSYS_CONFIG_BATTERY_CHARGER                (1)
 #define PBSYS_CONFIG_BLUETOOTH                      (1)
+#define PBSYS_CONFIG_BLUETOOTH_TOGGLE               (1)
+#define PBSYS_CONFIG_BLUETOOTH_TOGGLE_BUTTON        (512) // PBIO_BUTTON_RIGHT_UP
 #define PBSYS_CONFIG_HUB_LIGHT_MATRIX               (1)
 #define PBSYS_CONFIG_MAIN                           (1)
 #define PBSYS_CONFIG_PROGRAM_LOAD                   (1)