+++ lib/pbio/sys/hmi.c
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2018-2022 The Pybricks Authors
+// Copyright (c) 2018-2024 The Pybricks Authors
 
 // Provides Human Machine Interface (HMI) between hub and user.
 
-// TODO: implement additional buttons and Bluetooth light for SPIKE Prime
+// TODO: implement additional buttons and menu system (via matrix display) for SPIKE Prime
 // TODO: implement additional buttons and menu system (via screen) for NXT
 
 #include <stdbool.h>
@@ -19,6 +19,7 @@
 #include <pbio/color.h>
 #include <pbio/event.h>
 #include <pbio/light.h>
+#include <pbsys/bluetooth.h>
 #include <pbsys/config.h>
 #include <pbsys/status.h>
 
@@ -57,10 +58,45 @@ static PT_THREAD(update_program_run_button_wait_state(bool button_pressed)) {
     PT_END(pt);
 }
 
+#if PBSYS_CONFIG_BLUETOOTH_TOGGLE
+
+static struct pt update_bluetooth_button_wait_state_pt;
+
+/**
+ * Protothread to monitor the button state to toggle Bluetooth.
+ * @param [in]  button_pressed      The current button state.
+ */
+static PT_THREAD(update_bluetooth_button_wait_state(bool button_pressed)) {
+    struct pt *pt = &update_bluetooth_button_wait_state_pt;
+    // HACK: Misuse of protothread to reduce code size. This is the same
+    // as checking if the user program is running after each PT_WAIT.
+    if (pbsys_status_test(PBIO_PYBRICKS_STATUS_USER_PROGRAM_RUNNING)) {
+        goto start;
+    }
+
+    PT_BEGIN(pt);
+
+    for (;;) {
+    start:
+        // button may still be pressed during user program
+        PT_WAIT_UNTIL(pt, !button_pressed);
+        PT_WAIT_UNTIL(pt, button_pressed);
+        pbsys_bluetooth_enabled_state_request_toggle();
+    }
+
+    PT_END(pt);
+}
+
+#endif // PBSYS_CONFIG_BLUETOOTH_TOGGLE
+
 void pbsys_hmi_init(void) {
     pbsys_status_light_init();
     pbsys_hub_light_matrix_init();
     PT_INIT(&update_program_run_button_wait_state_pt);
+
+    #if PBSYS_CONFIG_BLUETOOTH_TOGGLE
+    PT_INIT(&update_bluetooth_button_wait_state_pt);
+    #endif // PBSYS_CONFIG_BLUETOOTH_TOGGLE
 }
 
 void pbsys_hmi_handle_event(process_event_t event, process_data_t data) {
@@ -101,6 +137,10 @@ void pbsys_hmi_poll(void) {
             pbsys_status_clear(PBIO_PYBRICKS_STATUS_POWER_BUTTON_PRESSED);
             update_program_run_button_wait_state(false);
         }
+
+        #if PBSYS_CONFIG_BLUETOOTH_TOGGLE
+        update_bluetooth_button_wait_state(btn & PBSYS_CONFIG_BLUETOOTH_TOGGLE_BUTTON);
+        #endif // PBSYS_CONFIG_BLUETOOTH_TOGGLE
     }
 
     pbsys_status_light_poll();